+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-...
+  * Configurable maximum post length
+  * Configuration for client sent an empty query language
 
 ### Fixed
 
-...
+  * Give an error for when a post is too long
 
 ### Changed